+++ persistent/Database/Persist/Sql/Orphan/PersistStore.hs
@@ -85,10 +85,10 @@ getTableName :: forall record m backend.
              ) => record -> ReaderT backend m Text
 getTableName rec = withReaderT projectBackend $ do
     conn <- ask
-    return $ connEscapeName conn $ tableDBName rec
+    return $ connEscapeTableName conn (entityDef $ Just rec)
 
 -- | useful for a backend to implement tableName by adding escaping
-tableDBName :: (PersistEntity record) => record -> DBName
+tableDBName :: (PersistEntity record) => record -> EntityNameDB
 tableDBName rec = entityDB $ entityDef (Just rec)
 
 -- | get the SQL string for the field that an EntityField represents
@@ -105,10 +105,10 @@ getFieldName :: forall record typ m backend.
              => EntityField record typ -> ReaderT backend m Text
 getFieldName rec = withReaderT projectBackend $ do
     conn <- ask
-    return $ connEscapeName conn $ fieldDBName rec
+    return $ connEscapeFieldName conn (fieldDB $ persistFieldDef rec)
 
 -- | useful for a backend to implement fieldName by adding escaping
-fieldDBName :: forall record typ. (PersistEntity record) => EntityField record typ -> DBName
+fieldDBName :: forall record typ. (PersistEntity record) => EntityField record typ -> FieldNameDB
 fieldDBName = fieldDB . persistFieldDef
 
 
@@ -143,7 +143,7 @@ instance PersistStoreWrite SqlBackend where
         let wher = whereStmtForKey conn k
         let sql = T.concat
                 [ "UPDATE "
-                , connEscapeName conn $ tableDBName $ recordTypeFromKey k
+                , connEscapeTableName conn (entityDef $ Just $ recordTypeFromKey k)
                 , " SET "
                 , T.intercalate "," $ map (mkUpdateText conn) upds
                 , " WHERE "
@@ -232,9 +232,9 @@ instance PersistStoreWrite SqlBackend where
           let valss = map mkInsertValues vals
           let sql = T.concat
                   [ "INSERT INTO "
-                  , connEscapeName conn (entityDB t)
+                  , connEscapeTableName conn t
                   , "("
-                  , T.intercalate "," $ map (connEscapeName conn . fieldDB) $ entityFields t
+                  , T.intercalate "," $ map (connEscapeFieldName conn . fieldDB) $ entityFields t
                   , ") VALUES ("
                   , T.intercalate "),(" $ replicate (length valss) $ T.intercalate "," $ map (const "?") (entityFields t)
                   , ")"
@@ -247,7 +247,7 @@ instance PersistStoreWrite SqlBackend where
         let wher = whereStmtForKey conn k
         let sql = T.concat
                 [ "UPDATE "
-                , connEscapeName conn (entityDB t)
+                , connEscapeTableName conn t
                 , " SET "
                 , T.intercalate "," (map (go conn . fieldDB) $ entityFields t)
                 , " WHERE "
@@ -256,7 +256,7 @@ instance PersistStoreWrite SqlBackend where
             vals = mkInsertValues val `mappend` keyToValues k
         rawExecute sql vals
       where
-        go conn x = connEscapeName conn x `T.append` "=?"
+        go conn x = connEscapeFieldName conn x `T.append` "=?"
 
     insertKey k v = insrepHelper "INSERT" [Entity k v]
 
@@ -296,7 +296,7 @@ instance PersistStoreWrite SqlBackend where
         wher conn = whereStmtForKey conn k
         sql conn = T.concat
             [ "DELETE FROM "
-            , connEscapeName conn $ tableDBName $ recordTypeFromKey k
+            , connEscapeTableName conn (entityDef $ Just $ recordTypeFromKey k)
             , " WHERE "
             , wher conn
             ]
@@ -328,7 +328,7 @@ instance PersistStoreRead SqlBackend where
                 [ "SELECT "
                 , cols conn
                 , " FROM "
-                , connEscapeName conn $ entityDB t
+                , connEscapeTableName conn t
                 , " WHERE "
                 , wher
                 ]
@@ -367,7 +367,7 @@ insrepHelper command es = do
     sql conn columnNames = T.concat
         [ command
         , " INTO "
-        , connEscapeName conn (entityDB entDef)
+        , connEscapeTableName conn entDef
         , "("
         , T.intercalate "," columnNames
         , ") VALUES ("